Move zap support to subpackage
Depending on bark shouldn't impose a dependency on zap.

This moves zap support into a zapify subpackage.
